test_imports.py
#!/usr/bin/env python3
"""
Simple import test to verify all dependencies work correctly
"""

import sys
import os

def test_imports():
    """Test all imports to ensure compatibility"""
    print("🧪 Testing imports...")
    
    try:
        # Test basic imports
        import streamlit as st
        print("✅ streamlit imported successfully")
        
        import json
        print("✅ json imported successfully")
        
        from datetime import datetime
        print("✅ datetime imported successfully")
        
        # Test LangChain imports
        from langchain.prompts import PromptTemplate, ChatPromptTemplate, SystemMessagePromptTemplate, HumanMessagePromptTemplate
        print("✅ langchain.prompts imported successfully")
        
        from langchain.chains import LLMChain, SequentialChain, ConversationChain
        print("✅ langchain.chains imported successfully")
        
        from langchain_openai import ChatOpenAI
        print("✅ langchain_openai imported successfully")
        
        from langchain.memory import ConversationBufferMemory, ConversationSummaryMemory
        print("✅ langchain.memory imported successfully")
        
        from langchain.agents import initialize_agent, AgentType, Tool
        print("✅ langchain.agents imported successfully")
        
        from langchain.output_parsers import PydanticOutputParser
        print("✅ langchain.output_parsers imported successfully")
        
        from langchain_community.callbacks import get_openai_callback
        print("✅ langchain_community.callbacks imported successfully")
        
        from pydantic import BaseModel, Field
        print("✅ pydantic imported successfully")
        
        from dotenv import load_dotenv
        print("✅ python-dotenv imported successfully")
        
        # Test our custom module
        import LangChainHelper
        print("✅ LangChainHelper imported successfully")
        
        print("\n🎉 All imports successful!")
        return True
        
    except ImportError as e:
        print(f"❌ Import error: {e}")
        return False
    except Exception as e:
        if "api_key" in str(e).lower() or "openai_api_key" in str(e).lower():
            print("✅ All imports successful! (API key not set - this is expected)")
            return True
        else:
            print(f"❌ Unexpected error: {e}")
            return False

def test_basic_functionality():
    """Test basic functionality without API calls"""
    print("\n🧪 Testing basic functionality...")
    
    try:
        from LangChainHelper import generate_store_name_and_items
        
        # Test without API key (should fail gracefully)
        result = generate_store_name_and_items("Basketball")
        print("✅ Basic function call successful")
        return True
        
    except Exception as e:
        print(f"⚠️  Function test failed (expected without API key): {e}")
        return True  # This is expected without API key

if __name__ == "__main__":
    print("🏀 Sports Store Analyzer - Import Test")
    print("=" * 40)
    
    # Test imports
    imports_ok = test_imports()
    
    if imports_ok:
        # Test basic functionality
        test_basic_functionality()
        
        print("\n✅ All tests completed!")
        print("🚀 You can now run: streamlit run main.py")
    else:
        print("\n❌ Import tests failed. Please check your dependencies.")
        print("💡 Try running: pip install -r requirements.txt") 