requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0 